Add USDZ model import; drop numbered panel eyebrows
Adds USDZ to the supported model formats via three's USDLoader (in-process parse, no world-axis rotation), wired through the file input, model bundle detection, and error messages.

Removes the / NN section numbers from panel eyebrows and widens the AO hemisphere kernel to Float64 for precision.

+++ src/lib/aoRaster.ts
@@ -194,7 +194,7 @@ function shadeAOTexel(
   w0: number,
   w1: number,
   w2: number,
-  kernel: Float32Array,
+  kernel: Float64Array,
   samples: number,
   bvh: MeshBVH,
   epsilon: number,
